SSBW_APIS/requirements.txt
Django>=3.0,<4.0
mongoengine
requests
djangorestframework
